requirements.txt
fastapi
uvicorn
jinja2
psycopg2-binary
python-multipart
numpy
pandas
requests
beautifulsoup4
lxml